Move StdFile
